Added 'label_for' key to $args when callin add_settings_field
Per the Wordpress codex for add_settings_field(...), the 'label_for' key
will be used to wrap a label tag around the setting title. with the
value as the corresponding input field id.  This change adds this key so
that fields using it will be labeled when rendered.  Also, the potential
callback functions are modified to use this additional array key for
consistency.